Add logic to save the srapted job details and allows to continue srapting after program stoped due to errors.
Adds job detail persistence and resumes scraping

Implements logic to save scraped job details to a file, allowing the scraper to resume after errors or interruptions. This prevents data loss and ensures more complete data collection.

Checks for existing job details before scraping, skipping already processed jobs.

+++ scripts/data_scraping/jobPos_scraper.py
@@ -4,6 +4,7 @@ import json
 import math
 import urllib
 import asyncio
+import pprint as pp
 
 from pathlib import Path
 from typing import Dict, List
@@ -12,13 +13,17 @@ from scrapfly import ScrapeApiResponse, ScrapeConfig, ScrapflyClient, ScrapflySc
 
 
 def parse_search_page(response):
-    html = re.findall(r"window.mosaic.providerData\[['\"]mosaic-provider-jobcards['\"]\]=(\{.+?\});", response.content)
-    html = json.loads(html[0])
-    job_dict = {
-        "job_results": html["metaData"]["mosaicProviderJobCardsModel"]["results"],
-        "search_meta": html["metaData"]["mosaicProviderJobCardsModel"]["tierSummaries"],
-    }
-    return job_dict
+    try:
+        html = re.findall(r"window.mosaic.providerData\[['\"]mosaic-provider-jobcards['\"]\]=(\{.+?\});", response.content)
+        html = json.loads(html[0])
+        job_dict = {
+            "job_results": html["metaData"]["mosaicProviderJobCardsModel"]["results"],
+                "search_meta": html["metaData"]["mosaicProviderJobCardsModel"]["tierSummaries"],
+            }
+        return job_dict
+    except Exception as e:
+        log.error(f"Failed to parse {response.config['url']}: {e}")
+        return None
 
 
 def add_url_params(url, **kwargs):
@@ -40,20 +45,20 @@ async def scrape_search(url: str, max_results: int = 1000) -> List[Dict]:
     
     if total_results > max_results:
         total_results = max_results
-    print(f"scraping remaining {(total_results - 10) / 10} pages")
+    print(f"Scraping remaining {(total_results - 10) / 10} pages...")
     
     other_pages = [
         ScrapeConfig(add_url_params(url, start=offset), **BASE_CONFIG)
         for offset in range(10, total_results + 10, 10)
     ]
-    log.info(f"found total pages {math.ceil(total_results / 10)} search pages")
+    log.info(f"found total {math.ceil(total_results / 10)} pages")
     
     async for result in SCRAPFLY.concurrent_scrape(other_pages):
         if not isinstance(result, ScrapflyScrapeError):
             data = parse_search_page(result)
             results.extend(data["job_results"])
         else:
-            log.error(f"failed to scrape {result.api_response.config['url']}, got: {result.message}")
+            log.error(f"Failed to scrape {result.api_response.config['url']}, got: {result.message}")
     return results
 
 
@@ -75,13 +80,17 @@ async def parse_job_page(response: ScrapeApiResponse):
 
 
 async def scrape_jobs(job_keys: List[str]):
-    """
-    Scrape job details and return both successful results and failed job keys
-    """
-    log.info(f"scraping {len(job_keys)} job listings")
-    successful_results = []
-    failed_job_keys = []
+    log.info(f"Scraping {len(job_keys)} job listings")
+    jobDetail_results = []
+    jobKeys_left = []
     job_key_to_url = {}
+    
+    if os.path.exists(JOB_DETAILS_FILE):
+        with open(JOB_DETAILS_FILE, "r") as f:
+            jobDetail_scraped = json.load(f)
+            jobKey_scraped = [job["jobkey"] for job in jobDetail_scraped]
+            jobKeys_left = [job for job in job_keys if job not in jobKey_scraped]
+        job_keys = jobKeys_left
     
     urls = []
     for job_key in job_keys:
@@ -92,52 +101,47 @@ async def scrape_jobs(job_keys: List[str]):
     to_scrape = [ScrapeConfig(url, **BASE_CONFIG) for url in urls]
     
     async for result in SCRAPFLY.concurrent_scrape(to_scrape):
-        job_detail = await parse_job_page(result)
-        if job_detail:
-            job_url = result.config["url"]
-            job_key = job_key_to_url[job_url]
-            job_detail["jobkey"] = job_key
-            successful_results.append(job_detail)
-        else:
-            failed_url = result.config["url"]
-            failed_job_key = job_key_to_url[failed_url]
-            failed_job_keys.append(failed_job_key)
-    
-    return {
-        "successful_results": successful_results,
-        "failed_job_keys": failed_job_keys
-    }
+        if result.result.get("result").get("success") == True:
+            job_detail = await parse_job_page(result)
+            
+            if job_detail:
+                job_url = result.config["url"]
+                job_key = job_key_to_url[job_url]
+                job_detail["jobkey"] = job_key
+                jobDetail_results.append(job_detail)
+                
+                with open(JOB_DETAILS_FILE, "w", encoding="utf-8") as f:
+                    json.dump(jobDetail_results, f, indent=4, ensure_ascii=False)
+        
+    log.info(f"Scraped {len(jobDetail_results)/len(job_keys)} job details successfully")
+    return jobDetail_results
 
 
 async def main():
-    url = "https://www.indeed.com/jobs?q=Data+Science%2C+Data+Scientist%2C+Data+Engineer"
-    search_results = await scrape_search(url, max_results=1000)
-    
-    with open(os.path.join(BASE_DIR, "data/processed_data/indeed_search_results.json"), "w", encoding="utf-8") as f:
-        json.dump(search_results, f, indent=4, ensure_ascii=False)
-    
-    job_keys = [job["jobkey"] for job in search_results if "jobkey" in job]
-    
-    with open(os.path.join(BASE_DIR, "data/processed_data/indeed_jobKeys.json"), "w", encoding="utf-8") as f:
-        json.dump(job_keys, f, indent=4, ensure_ascii=False)
+    if not os.path.exists(SEARCH_RESULTS_FILE):
+        url = "https://www.indeed.com/jobs?q=Data%20Science"
+        search_results = await scrape_search(url)
+        
+        with open(SEARCH_RESULTS_FILE, "w", encoding="utf-8") as f:
+            json.dump(search_results, f, indent=4, ensure_ascii=False)
+        
+        job_keys = [job["jobkey"] for job in search_results if "jobkey" in job]
+            
+        with open(JOB_KEYS_FILE, "w", encoding="utf-8") as f:
+            json.dump(job_keys, f, indent=4, ensure_ascii=False)
+    else:
+        with open(JOB_KEYS_FILE, "r", encoding="utf-8") as f:
+            job_keys = json.load(f)
+        
+        with open(SEARCH_RESULTS_FILE, "r", encoding="utf-8") as f:
+            search_results = json.load(f)
     
     if job_keys:
-        scrape_result = await scrape_jobs(job_keys)
-        job_details = scrape_result["successful_results"]
-        failed_keys = scrape_result["failed_job_keys"]
-        
-        if job_details:
-            with open(os.path.join(BASE_DIR, "data/processed_data/indeed_job_details.json"), "w", encoding="utf-8") as f:
-                json.dump(job_details, f, indent=4, ensure_ascii=False)
-        
-        if failed_keys:
-            log.warning(f"Failed to scrape {len(failed_keys)} jobs. Saving for later retry.")
-            with open(os.path.join(BASE_DIR, "data/processed_data/indeed_failed_jobKeys.json"), "w", encoding="utf-8") as f:
-                json.dump(failed_keys, f, indent=4, ensure_ascii=False)
+        jobDetail_results = await scrape_jobs(job_keys)
         
         job_details_dict = {
             detail.get("jobkey"): detail
-            for detail in job_details
+            for detail in jobDetail_results
             if detail.get("jobkey")
         }
         
@@ -145,21 +149,28 @@ async def main():
             key = job.get("jobkey")
             if key and key in job_details_dict:
                 job.update(job_details_dict[key])
-    
-    with open(os.path.join(BASE_DIR, "data/processed_data/indeed_jobs_final.json"), "w", encoding="utf-8") as f:
+        
+    with open(JOBS_FINAL_FILE, "w", encoding="utf-8") as f:
         json.dump(search_results, f, indent=4, ensure_ascii=False)
 
 
 if __name__ == "__main__":
+    #* File path constants
     BASE_DIR = Path(__file__).resolve().parents[2]
+    JOB_KEYS_FILE = os.path.join(BASE_DIR, "data/raw_data/data_science/indeed_jobKeys.json")
+    SEARCH_RESULTS_FILE = os.path.join(BASE_DIR, "data/raw_data/data_science/indeed_search_results.json")
+    JOB_DETAILS_FILE = os.path.join(BASE_DIR, "data/raw_data/data_science/indeed_job_details.json")
+    FAILED_JOB_KEYS_FILE = os.path.join(BASE_DIR, "data/raw_data/data_science/indeed_failed_jobKeys.json")
+    JOBS_FINAL_FILE = os.path.join(BASE_DIR, "data/raw_data/data_science/indeed_jobs_final.json")
+    
     SCRAPFLY = ScrapflyClient(key=os.environ["SCRAPFLY_API_KEY"])
     BASE_CONFIG = {
         "asp": True,
         "cost_budget": 1000,
         "country": "US",
         "render_js": True,
+        "rendering_wait": 2000,
         "proxy_pool": "public_residential_pool",
-        "debug": True,
     }
     
     asyncio.run(main())
